requirements.txt
dropbox==11.36.2
numpy==1.24.1
Pillow==10.1.0
python-decouple==3.8
Requests==2.32.3
